pyproject.toml
[project]
name = "flarecrawl"
version = "0.11.0"
description = "Cloudflare Browser Rendering CLI - Firecrawl-compatible, cost-efficient at scale"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Fabric Team" }
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx[http2]>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
flarecrawl = "flarecrawl.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flarecrawl"]

# ============================================================
# FABRIC METADATA (Required for registry)
# ============================================================
[tool.clique]
description = "Cloudflare Browser Rendering - scrape, crawl, map, download, extract, screenshot, pdf, favicon"
resources = ["scrape", "crawl", "map", "download", "extract", "screenshot", "pdf", "favicon", "schema", "discover"]
auth = "api_key"
status = "experimental"
origin = "clique"
env-key = "FLARECRAWL_API_TOKEN"
api-url = "https://dash.cloudflare.com/"
system-deps = []

# ============================================================
# CODE QUALITY
# ============================================================
[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
